+++ src/main/java/com/incountry/residence/sdk/dto/search/FilterNumberParam.java
@@ -25,17 +25,17 @@ public class FilterNumberParam {
     private static final String ERR_CONDITION_RESTR = String.format("Operator in number filter can by only in [%s,%s,%s,%s,%s]",
             OPER_LT, OPER_LTE, OPER_GT, OPER_GTE, OPER_NOT);
 
-    private final Integer[] values;
+    private final Long[] values;
     private final String operator1;
     private final String operator2;
 
-    private FilterNumberParam(Integer[] values, String operator1, String operator2) {
+    private FilterNumberParam(Long[] values, String operator1, String operator2) {
         this.values = Arrays.copyOf(values, values.length);
         this.operator1 = operator1;
         this.operator2 = operator2;
     }
 
-    public FilterNumberParam(Integer[] values) throws StorageClientException {
+    public FilterNumberParam(Long[] values) throws StorageClientException {
         if (values == null || values.length == 0 || (Stream.of(values).anyMatch(Objects::isNull))) {
             LOG.error(ERR_NULL_VALUE);
             throw new StorageClientException(ERR_NULL_VALUE);
@@ -45,7 +45,7 @@ public class FilterNumberParam {
         this.operator2 = null;
     }
 
-    public FilterNumberParam(String operator, Integer value) throws StorageClientException {
+    public FilterNumberParam(String operator, Long value) throws StorageClientException {
         if (operator == null || notIn(operator, OPER_GT, OPER_GTE, OPER_NOT, OPER_LT, OPER_LTE)) {
             LOG.error(ERR_CONDITION_RESTR);
             throw new StorageClientException(ERR_CONDITION_RESTR);
@@ -54,12 +54,12 @@ public class FilterNumberParam {
             LOG.error(ERR_NULL_VALUE);
             throw new StorageClientException(ERR_NULL_VALUE);
         }
-        this.values = new Integer[]{value};
+        this.values = new Long[]{value};
         this.operator1 = operator;
         this.operator2 = null;
     }
 
-    public FilterNumberParam(String operator1, Integer value1, String operator2, Integer value2) throws StorageClientException {
+    public FilterNumberParam(String operator1, Long value1, String operator2, Long value2) throws StorageClientException {
         if (operator1 == null || notIn(operator1, OPER_GT, OPER_GTE)) {
             LOG.error(ERR_OPER1_RESTR);
             throw new StorageClientException(ERR_OPER1_RESTR);
@@ -76,7 +76,7 @@ public class FilterNumberParam {
             LOG.error(ERR_RANGE_RESTR);
             throw new StorageClientException(ERR_RANGE_RESTR);
         }
-        this.values = new Integer[]{value1, value2};
+        this.values = new Long[]{value1, value2};
         this.operator1 = operator1;
         this.operator2 = operator2;
     }
@@ -99,11 +99,11 @@ public class FilterNumberParam {
         return operator2 != null && operator1 != null;
     }
 
-    public Integer[] getValues() {
+    public Long[] getValues() {
         if (values != null && values.length > 0) {
             return Arrays.copyOf(values, values.length);
         }
-        return new Integer[]{};
+        return new Long[]{};
     }
 
     public String getOperator1() {
